requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
chromadb>=0.5.0
sqlalchemy>=2.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
pymupdf>=1.24.0
pillow>=10.3.0
pytesseract>=0.3.10
pandas>=2.2.0
python-multipart>=0.0.9
structlog>=24.2.0
tenacity>=8.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
python-docx>=1.1.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0
